SetClassByPosition: Also replace old class in CNWLevelStats
While NWNX_Creature_SetClassByPosition correctly replaced the Class ID in the ClassList node, it did not change references to the old class in LvlStatList. This means that when trying to level up after having used the function, the game got confused, would throw a "[Server] Invalid number of multiclasses" error message and revert the levelup.

This commit modifies the function's default behavior to also update these old class references. Should this not be desired, the old functionality remains accessible via the bUpdateLevels argument.